cli/pyproject.toml
[project]
name = "cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.0.0",
    "pyobjc-framework-cocoa>=12.1",
    "pyobjc-framework-quartz>=12.1",
    "pyobjc-framework-screencapturekit>=12.1",
]
